+++ firmware/src/led.h
@@ -4,8 +4,8 @@
 #define LED_H
 #include <stdint.h>
 
-int led_init(void);
-void led_setColor(uint32_t red, uint32_t green, uint32_t blue);
+int8_t led_init(void);
+void led_setColor(uint8_t red, uint8_t green, uint8_t blue);
 void led_on(void);
 void led_off(void);
